+++ cmd/influx/pkg.go
@@ -8,6 +8,7 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
+	"net/url"
 	"os"
 	"path"
 	"path/filepath"
@@ -142,7 +143,7 @@ func (b *cmdPkgBuilder) cmdPkgApply() *cobra.Command {
 			-f $PATH_TO_TEMPLATE/template_2.yml
 
 		# Apply a template from a url
-		influx apply -u https://raw.githubusercontent.com/influxdata/community-templates/master/docker/docker.yml
+		influx apply -f https://raw.githubusercontent.com/influxdata/community-templates/master/docker/docker.yml
 
 		# Apply a template from STDIN
 		cat $TEMPLATE.json | influx apply --encoding json
@@ -314,8 +315,6 @@ func (b *cmdPkgBuilder) pkgExportRunEFn(cmd *cobra.Command, args []string) error
 		return err
 	}
 
-	opts := []pkger.CreatePkgSetFn{}
-
 	resTypes := []struct {
 		kind   pkger.Kind
 		idStrs []string
@@ -330,6 +329,8 @@ func (b *cmdPkgBuilder) pkgExportRunEFn(cmd *cobra.Command, args []string) error
 		{kind: pkger.KindTelegraf, idStrs: strings.Split(b.exportOpts.telegrafs, ",")},
 		{kind: pkger.KindVariable, idStrs: strings.Split(b.exportOpts.variables, ",")},
 	}
+
+	var opts []pkger.CreatePkgSetFn
 	for _, rt := range resTypes {
 		newOpt, err := newResourcesToClone(rt.kind, rt.idStrs)
 		if err != nil {
@@ -342,9 +343,17 @@ func (b *cmdPkgBuilder) pkgExportRunEFn(cmd *cobra.Command, args []string) error
 		return b.exportPkg(cmd.OutOrStdout(), pkgSVC, b.file, opts...)
 	}
 
-	kind := pkger.Kind(b.exportOpts.resourceType)
-	if err := kind.OK(); err != nil {
-		return errors.New("resource type must be one of bucket|dashboard|label|variable; got: " + b.exportOpts.resourceType)
+	resType := strings.ToLower(b.exportOpts.resourceType)
+	resKind := pkger.KindUnknown
+	for _, k := range pkger.Kinds() {
+		if strings.ToLower(string(k)) == resType {
+			resKind = k
+			break
+		}
+	}
+
+	if err := resKind.OK(); err != nil {
+		return errors.New("resource type is invalid; got: " + b.exportOpts.resourceType)
 	}
 
 	if stdin, err := b.inStdIn(); err == nil {
@@ -354,7 +363,7 @@ func (b *cmdPkgBuilder) pkgExportRunEFn(cmd *cobra.Command, args []string) error
 		}
 	}
 
-	resTypeOpt, err := newResourcesToClone(kind, args)
+	resTypeOpt, err := newResourcesToClone(resKind, args)
 	if err != nil {
 		return err
 	}
@@ -699,7 +708,7 @@ func (b *cmdPkgBuilder) stackListRunEFn(cmd *cobra.Command, args []string) error
 	defer tabW.Flush()
 
 	tabW.HideHeaders(b.hideHeaders)
-	tabW.WriteHeaders("ID", "OrgID", "Name", "Description", "Num Resources", "URLs", "Created At")
+	tabW.WriteHeaders("ID", "OrgID", "Name", "Description", "Num Resources", "Sources", "URLs", "Created At")
 
 	for _, stack := range stacks {
 		tabW.Write(map[string]interface{}{
@@ -708,6 +717,7 @@ func (b *cmdPkgBuilder) stackListRunEFn(cmd *cobra.Command, args []string) error
 			"Name":          stack.Name,
 			"Description":   stack.Description,
 			"Num Resources": len(stack.Resources),
+			"Sources":       stack.Sources,
 			"URLs":          stack.URLs,
 			"Created At":    stack.CreatedAt,
 		})
@@ -816,11 +826,12 @@ func (b *cmdPkgBuilder) registerPkgPrintOpts(cmd *cobra.Command) {
 }
 
 func (b *cmdPkgBuilder) registerPkgFileFlags(cmd *cobra.Command) {
-	cmd.Flags().StringSliceVarP(&b.files, "file", "f", nil, "Path to template file")
+	cmd.Flags().StringSliceVarP(&b.files, "file", "f", nil, "Path to template file; Supports HTTP(S) URLs or file paths.")
 	cmd.MarkFlagFilename("file", "yaml", "yml", "json", "jsonnet")
 	cmd.Flags().BoolVarP(&b.recurse, "recurse", "R", false, "Process the directory used in -f, --file recursively. Useful when you want to manage related templates organized within the same directory.")
 
 	cmd.Flags().StringSliceVarP(&b.urls, "template-url", "u", nil, "URL to template file")
+	cmd.Flags().MarkHidden("template-url")
 
 	cmd.Flags().StringVarP(&b.encoding, "encoding", "e", "", "Encoding for the input stream. If a file is provided will gather encoding type from file extension. If extension provided will override.")
 	cmd.MarkFlagFilename("encoding", "yaml", "yml", "json", "jsonnet")
@@ -891,12 +902,25 @@ func (b *cmdPkgBuilder) readRawPkgsFromURLs(urls []string) ([]*pkger.Pkg, error)
 }
 
 func (b *cmdPkgBuilder) readPkg() (*pkger.Pkg, bool, error) {
-	pkgs, err := b.readRawPkgsFromFiles(b.files, b.recurse)
+	var remotes, files []string
+	for _, rawURL := range append(b.files, b.urls...) {
+		u, err := url.Parse(rawURL)
+		if err != nil {
+			return nil, false, ierror.Wrap(err, fmt.Sprintf("failed to parse url[%s]", rawURL))
+		}
+		if strings.HasPrefix(u.Scheme, "http") {
+			remotes = append(remotes, u.String())
+		} else {
+			files = append(files, u.String())
+		}
+	}
+
+	pkgs, err := b.readRawPkgsFromFiles(files, b.recurse)
 	if err != nil {
 		return nil, false, err
 	}
 
-	urlPkgs, err := b.readRawPkgsFromURLs(b.urls)
+	urlPkgs, err := b.readRawPkgsFromURLs(remotes)
 	if err != nil {
 		return nil, false, err
 	}