+++ cmd/influxd/launcher/launcher_helpers.go
@@ -51,11 +51,12 @@ type TestLauncher struct {
 }
 
 // NewTestLauncher returns a new instance of TestLauncher.
-func NewTestLauncher() *TestLauncher {
+func NewTestLauncher(flagger feature.Flagger) *TestLauncher {
 	l := &TestLauncher{Launcher: NewLauncher()}
 	l.Launcher.Stdin = &l.Stdin
 	l.Launcher.Stdout = &l.Stdout
 	l.Launcher.Stderr = &l.Stderr
+	l.Launcher.flagger = flagger
 	if testing.Verbose() {
 		l.Launcher.Stdout = io.MultiWriter(l.Launcher.Stdout, os.Stdout)
 		l.Launcher.Stderr = io.MultiWriter(l.Launcher.Stderr, os.Stderr)
@@ -70,9 +71,9 @@ func NewTestLauncher() *TestLauncher {
 }
 
 // RunTestLauncherOrFail initializes and starts the server.
-func RunTestLauncherOrFail(tb testing.TB, ctx context.Context, args ...string) *TestLauncher {
+func RunTestLauncherOrFail(tb testing.TB, ctx context.Context, flagger feature.Flagger, args ...string) *TestLauncher {
 	tb.Helper()
-	l := NewTestLauncher()
+	l := NewTestLauncher(flagger)
 
 	if err := l.Run(ctx, args...); err != nil {
 		tb.Fatal(err)