+++ pkger/service_logging.go
@@ -101,6 +101,45 @@ func (s *loggingMW) ListStacks(ctx context.Context, orgID influxdb.ID, f ListFil
 	return s.next.ListStacks(ctx, orgID, f)
 }
 
+func (s *loggingMW) ReadStack(ctx context.Context, id influxdb.ID) (st Stack, err error) {
+	defer func(start time.Time) {
+		if err != nil {
+			s.logger.Error("failed to read stack",
+				zap.Error(err),
+				zap.String("id", id.String()),
+				zap.Duration("took", time.Since(start)),
+			)
+			return
+		}
+	}(time.Now())
+	return s.next.ReadStack(ctx, id)
+}
+
+func (s *loggingMW) UpdateStack(ctx context.Context, upd StackUpdate) (_ Stack, err error) {
+	defer func(start time.Time) {
+		if err != nil {
+			fields := []zap.Field{
+				zap.Error(err),
+				zap.String("id", upd.ID.String()),
+			}
+			if upd.Name != nil {
+				fields = append(fields, zap.String("name", *upd.Name))
+			}
+			if upd.Description != nil {
+				fields = append(fields, zap.String("desc", *upd.Description))
+			}
+			fields = append(fields,
+				zap.Strings("urls", upd.URLs),
+				zap.Duration("took", time.Since(start)),
+			)
+
+			s.logger.Error("failed to update stack", fields...)
+			return
+		}
+	}(time.Now())
+	return s.next.UpdateStack(ctx, upd)
+}
+
 func (s *loggingMW) CreatePkg(ctx context.Context, setters ...CreatePkgSetFn) (pkg *Pkg, err error) {
 	defer func(start time.Time) {
 		dur := zap.Duration("took", time.Since(start))