+++ ui/src/shared/components/inlineLabels/InlineLabelsEditor.tsx
@@ -9,6 +9,7 @@ import {
   ButtonShape,
   ButtonBaseRef,
   ComponentColor,
+  ClickOutside,
 } from '@influxdata/clockface'
 import InlineLabelPopover from 'src/shared/components/inlineLabels/InlineLabelPopover'
 import CreateLabelOverlay from 'src/labels/components/CreateLabelOverlay'
@@ -53,6 +54,7 @@ interface State {
   searchTerm: string
   selectedItemID: string
   isCreatingLabel: OverlayState
+  isPopoverVisible: boolean
 }
 
 @ErrorHandling
@@ -66,7 +68,12 @@ class InlineLabelsEditor extends Component<Props, State> {
       selectedItemID: null,
       searchTerm: '',
       isCreatingLabel: OverlayState.Closed,
+      isPopoverVisible: false,
     }
+  }
+
+  public componentDidMount() {
+    this.handleAddPopoverEventListener()
   }
 
   public render() {
@@ -101,23 +108,26 @@ class InlineLabelsEditor extends Component<Props, State> {
 
   private get popover(): JSX.Element {
     const {labels, selectedLabels} = this.props
-    const {searchTerm, selectedItemID} = this.state
+    const {searchTerm, selectedItemID, isPopoverVisible} = this.state
 
     const labelsUsed =
       labels.length > 0 && labels.length === selectedLabels.length
 
     return (
-      <InlineLabelPopover
-        searchTerm={searchTerm}
-        triggerRef={this.popoverTrigger}
-        selectedItemID={selectedItemID}
-        onUpdateSelectedItemID={this.handleUpdateSelectedItemID}
-        allLabelsUsed={labelsUsed}
-        onStartCreatingLabel={this.handleStartCreatingLabel}
-        onInputChange={this.handleInputChange}
-        filteredLabels={this.filterLabels(searchTerm)}
-        onAddLabel={this.handleAddLabel}
-      />
+      <ClickOutside onClickOutside={this.onClickOutside}>
+        <InlineLabelPopover
+          searchTerm={searchTerm}
+          triggerRef={this.popoverTrigger}
+          selectedItemID={selectedItemID}
+          onUpdateSelectedItemID={this.handleUpdateSelectedItemID}
+          allLabelsUsed={labelsUsed}
+          onStartCreatingLabel={this.handleStartCreatingLabel}
+          onInputChange={this.handleInputChange}
+          filteredLabels={this.filterLabels(searchTerm)}
+          onAddLabel={this.handleAddLabel}
+          visible={isPopoverVisible}
+        />
+      </ClickOutside>
     )
   }
 
@@ -143,6 +153,27 @@ class InlineLabelsEditor extends Component<Props, State> {
         <span className="cf-label--name">Add a label</span>
       </div>
     )
+  }
+
+  private handleAddPopoverEventListener = (): void => {
+    if (!this.popoverTrigger.current) {
+      return
+    }
+    this.popoverTrigger.current.addEventListener('click', () => {
+      this.setState({
+        isPopoverVisible: true,
+      })
+    })
+  }
+
+  private onClickOutside = (e: MouseEvent): void => {
+    if (e.target === this.popoverTrigger.current) {
+      return
+    }
+
+    this.setState({
+      isPopoverVisible: false,
+    })
   }
 
   private handleAddLabel = async (labelID: string) => {
@@ -243,7 +274,7 @@ class InlineLabelsEditor extends Component<Props, State> {
   }
 
   private handleStartCreatingLabel = (): void => {
-    this.setState({isCreatingLabel: OverlayState.Open})
+    this.setState({isCreatingLabel: OverlayState.Open, isPopoverVisible: false})
   }
 
   private handleStopCreatingLabel = (): void => {