+++ ui/src/shared/utils/formatStatValue.ts
@@ -48,7 +48,7 @@ export const formatStatValue = (
           maximumFractionDigits: MAX_DECIMAL_PLACES,
         })
   } else if (isString(value)) {
-    localeFormattedValue = value ? Number(value).toFixed(digits) : value
+    localeFormattedValue = value
   } else {
     return 'Data cannot be displayed'
   }